create the dto's to response and request to BookController
