+++ Library'sProject/src/main/java/com/riwi/library/application/IModels/IModelAutor.java
@@ -1,7 +1,7 @@
 package com.riwi.library.application.IModels;
 
-import com.riwi.library.infrastructure.dto.request.AutorCreateRequestDTO;
-import com.riwi.library.infrastructure.dto.response.AutorCreateResponseDTO;
+import com.riwi.library.infrastructure.dto.request.autorRequestDtos.AutorCreateRequestDTO;
+import com.riwi.library.infrastructure.dto.response.autorResponseDtos.AutorCreateResponseDTO;
 
 public interface IModelAutor {
     AutorCreateResponseDTO create(AutorCreateRequestDTO autorCreateRequestDTO);
