+++ Library'sProject/src/main/java/com/riwi/library/application/services/BookService.java~
@@ -25,9 +25,6 @@ public class BookService implements IModelBook {
                 .build();
 
         bookRepository.save(book);
-        return BookCreateResponseDTO.builder()
-                .title(book.getTitle())
-                .author(book.getAuthor())
-                .build();
+        return book.toString();
     }
 }
