+++ Library'sProject/src/main/java/com/riwi/library/infrastructure/IModels/IBookModel.java
@@ -1,8 +1,10 @@
 package com.riwi.library.infrastructure.IModels;
 
+import com.riwi.library.infrastructure.dto.request.bookRequestDtos.BookCreateRequestDTO;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
 
 public interface IBookModel {
 
-    ResponseEntity<?> create(String title, String author, String year, String description);
+    ResponseEntity<?> create(@RequestBody BookCreateRequestDTO bookCreateRequestDTO);
 }
