+++ Library'sProject/src/main/java/com/riwi/library/infrastructure/controllers/BookController.java
@@ -2,10 +2,12 @@ package com.riwi.library.infrastructure.controllers;
 
 import com.riwi.library.application.services.BookService;
 import com.riwi.library.infrastructure.IModels.IBookModel;
+import com.riwi.library.infrastructure.dto.request.bookRequestDtos.BookCreateRequestDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -18,9 +20,9 @@ public class BookController implements IBookModel {
 
     @Override
     @PostMapping("/create")
-    public ResponseEntity<?> create(String title, String author, String year, String description) {
+    public ResponseEntity<?> create(@RequestBody BookCreateRequestDTO bookCreateRequestDTO) {
         try{
-            return ResponseEntity.status(HttpStatus.CREATED).body(bookService.create(title, author, year, description));
+            return ResponseEntity.status(HttpStatus.CREATED).body(bookService.create(bookCreateRequestDTO));
         }catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
