+++ Library'sProject/src/main/java/com/riwi/library/infrastructure/dto/request/bookRequestDtos/BookCreateRequestDTO.java
@@ -1,0 +1,17 @@
+package com.riwi.library.infrastructure.dto.request.bookRequestDtos;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class BookCreateRequestDTO {
+
+    private String title;
+    private String author;
+    private String year;
+    private String description;
+}
