--- Library'sProject/src/main/java/com/riwi/library/infrastructure/dto/response/AutorCreateResponseDTO.java
@@ -1,15 +1,0 @@
-package com.riwi.library.infrastructure.dto.response;
-
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class AutorCreateResponseDTO {
-
-    private String name;
-    private String description;
-}
