+++ Library'sProject/src/main/java/com/riwi/library/infrastructure/dto/response/bookResponseDtos/BookCreateResponseDTO.java
@@ -1,0 +1,15 @@
+package com.riwi.library.infrastructure.dto.response.bookResponseDtos;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class BookCreateResponseDTO {
+
+    private String title;
+    private String author;
+}
